comments and try-catch blocks

+++ CASWebApi/Controllers/TeacherController.cs
@@ -33,12 +33,18 @@ namespace CASWebApi.Controllers
         public  ActionResult<List<Teacher>> GetAllTeachers() 
         {
             logger.LogInformation("Getting all Teachers from TeacherController");
-            var teacherList = _teacherService.GetAll();
-            if (teacherList != null)
+            try
+            {
+                var teacherList = _teacherService.GetAll();
                 logger.LogInformation("Fetched All teacher data");
-            else
-                logger.LogError("Cannot get access to teacher collection in Db");
-            return teacherList;
+                return teacherList;
+            }
+            catch (Exception e)
+            {
+                logger.LogError("Cannot get access to Db");
+                return StatusCode(500, "Internal server error");
+            }
+
         }
 
         /// <summary>
@@ -53,15 +59,23 @@ namespace CASWebApi.Controllers
         public ActionResult<Teacher> GetTeacherById(string id)
         {
             logger.LogInformation("Getting Teacher by given Id from TeacherController");
-            var teacher = _teacherService.GetById(id);
-
-            if (teacher == null)
+            try
             {
-                logger.LogError("Cannot get access to teacher collection in Db");
-            }
-            logger.LogInformation("Fetched teacher data by id");
+                var teacher = _teacherService.GetById(id);
+                if (teacher == null)
+                {
+                    logger.LogError("Cannot get access to teacher collection in Db");
+                }
+                logger.LogInformation("Fetched teacher data by id");
 
-            return teacher;
+                return teacher;
+            }
+            catch(Exception e)
+            {
+                logger.LogError("Cannot get access to Db");
+                return StatusCode(500, "Internal server error");
+            }
+          
         }
 
         /// <summary>
@@ -73,15 +87,22 @@ namespace CASWebApi.Controllers
         public ActionResult<List<Teacher>> GetTeachersByCourseName(string courseName)
         {
             logger.LogInformation("Getting Teacher by given courseName from teacherController");
-            var teacherList = _teacherService.GetTeachersByCourseName(courseName);
-
-            if (teacherList == null)
+            try
             {
-                logger.LogError("Cannot get access to teacher collection in Db");
+                var teacherList = _teacherService.GetTeachersByCourseName(courseName);
+                if (teacherList == null)
+                {
+                    logger.LogError("Cannot get access to teacher collection in Db");
+                }
+                logger.LogInformation("Fetched teacher data by courseName");
+                return teacherList;
             }
-            logger.LogInformation("Fetched teacher data by courseName");
-
-            return teacherList;
+            catch(Exception e)
+            {
+                logger.LogError("Cannot get access to Db");
+                return StatusCode(500, "Internal server error");
+            }
+           
         }
 
         /// <summary>
@@ -92,24 +113,23 @@ namespace CASWebApi.Controllers
         [HttpPost("createTeacher", Name = nameof(CreateTeacher))]
         public ActionResult<Teacher> CreateTeacher(Teacher teacher)
         {
+            logger.LogInformation("Creating a new teacher profile: " + teacher);
             if (teacher != null)
             {
-                
-                logger.LogInformation("Creating a new teacher profile: " + teacher);
-                if (_teacherService.Create(teacher))
+                try
                 {
-                    logger.LogInformation("A new teacher profile added successfully " + teacher);
+                    _teacherService.Create(teacher);
                     return CreatedAtRoute("getTeacherById", new { id = teacher.Id }, teacher);
-                }
-                else
+                }              
+                catch(Exception e)
                 {
                     logger.LogError("Cannot create a teacher, duplicated id or wrong format");
-                    return NotFound("duplicated id or wrong id format");
+                    return Conflict(e);
                 }
 
             }
-
-            return BadRequest(false);
+            logger.LogError("A new Teacher model is not valid");
+            return BadRequest("Teacher param or his Id is null ");
         }
 
         /// <summary>
@@ -121,20 +141,31 @@ namespace CASWebApi.Controllers
         public IActionResult UpdateTeacher(Teacher teacherIn)
         {
             logger.LogInformation("Updating existed teacher profile: " +teacherIn.Id);
-            var teacher = _teacherService.GetById(teacherIn.Id);
-
-            if (teacher == null)
+            if(teacherIn != null && teacherIn.Id != null)
             {
-                logger.LogError("Teacher with Id: " + teacherIn.Id + " doesn't exist");
-                return NotFound(false);
-            }
-            bool updated = _teacherService.Update(teacherIn.Id, teacherIn);
-            if (updated)
-                logger.LogInformation("Given Teacher profile Updated successfully");
-            else
-                logger.LogError("Cannot update the teacher profile: "+ teacherIn.Id+" wrong format");
+                try
+                {
+                    var teacher = _teacherService.GetById(teacherIn.Id);
+                    if (teacher == null)
+                    {
+                        logger.LogError("Teacher with Id: " + teacherIn.Id + " doesn't exist");
+                        return NotFound(false);
+                    }
+                    bool updated = _teacherService.Update(teacherIn.Id, teacherIn);
+                    if (updated)
+                        logger.LogInformation("Given Teacher profile Updated successfully");
+                    else
+                        logger.LogError("Cannot update the teacher profile: " + teacherIn.Id + " wrong format");
 
-            return Ok(updated);
+                    return Ok(updated);
+                }
+                catch (Exception e)
+                {
+                    return BadRequest(e);
+                }
+            }
+            logger.LogError("A given Teacher model is not valid");
+            return BadRequest("Teacher param or Id is null ");
         }
 
         /// <summary>
@@ -145,16 +176,26 @@ namespace CASWebApi.Controllers
         [HttpDelete("deleteTeacherById", Name = nameof(DeleteTeacherById))]
         public IActionResult DeleteTeacherById(string id)
         {
-            logger.LogInformation("Deleting existed teacher profile: " + id);
-            var teacher = _teacherService.GetById(id);
-
-            if (teacher != null && _teacherService.RemoveById(teacher.Id))
+            if(id != null)
             {
-                logger.LogInformation("Teacher profile has been deleted successfully " + teacher);
-                return Ok(true);
+                try
+                {
+                    logger.LogInformation("Deleting existed teacher profile: " + id);
+                    var teacher = _teacherService.GetById(id);
+
+                    if (teacher != null && _teacherService.RemoveById(teacher.Id))
+                    {
+                        logger.LogInformation("Teacher profile has been deleted successfully " + teacher);
+                        return Ok(true);
+                    }
+                }
+                catch (Exception e)
+                {
+                    return BadRequest("No connection to database");
+                }
             }
-            logger.LogError("Teacher not found in Database");   
-            return NotFound(false);
+            logger.LogError("Id is not valid or empty string");
+            return NotFound("Id is not valid or empty string");
         }
     }
 }
